fix: jwtService.js move directory
