+++ server/src/services/jwtService.js
@@ -1,5 +1,5 @@
 import jwt from 'jsonwebtoken'
-import config from '../../config/config.js'
+import config from '../config/config.js'
 
 const secret = config.jwtSecret;
